nodedevice: Use the inverse of ctm as child ctm for path clips
We've previously just used ctm itself, which just happend to work for
the PDFs I've tested this with. Ideally, we could just transform the
path by the ctm instead of having to invert the matrix, but that isn't
possible.

It might however be worth to replace the matrix invert from mupdf with
the transform invert from GSK, as it can be faster for some
transformations.